backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
SQLAlchemy==2.0.48
python-dotenv==1.2.2
requests==2.32.5
firebase-admin==7.0.0
alembic==1.16.5
